requirements.txt
pytest==7.4.2
selenium==4.13.0
pytest-xdist==2.1.0
allure-pytest==2.13.2
